+++ jarboot-core/src/main/java/com/mz/jarboot/core/constant/CoreConstant.java
@@ -8,7 +8,6 @@ import java.io.File;
  */
 public class CoreConstant {
     public static final String EMPTY_STRING = "";
-    public static final String NULL_STRING = "null";
     public static final String LOG_NAME = "jarboot-core";
     public static final String BR = "<br/>";